requirements.txt
cerebras_cloud_sdk
fastapi
uvicorn
pydantic
requests
httpx
pinecone
sentence-transformers
torch
transformers
z3-solver
streamlit
jinja2
weasyprint
pdfkit
python-dotenv
numpy
PyPDF2